+++ clodoo/examples/import_products.py
@@ -56,7 +56,7 @@ except ImportError:
 # import pdb
 
 
-__version__ = "2.0.6"
+__version__ = "2.0.7"
 
 
 msg_time = time.time()