+++ clodoo/moduli_da_installare.py
@@ -20,7 +20,7 @@ except ImportError:
 # import transodoo
 # import pdb
 
-__version__ = "2.0.6"
+__version__ = "2.0.7"
 
 
 VERSIONS = ["vg7", "61", "70", "80", "90", "100", "110", "120"]