+++ clodoo/tests/__clodoo_test_02.py
@@ -26,7 +26,7 @@ except ImportError:
     import clodoo
 
 
-__version__ = "2.0.6"
+__version__ = "2.0.7"
 
 
 MODULE_ID = "clodoo"