+++ clodoo/tests/__pytest_transodoo.py
@@ -12,7 +12,7 @@
 # import pytest
 from clodoo import transodoo
 
-__version__ = "2.0.6"
+__version__ = "2.0.7"
 
 
 class TestClass: