+++ wok_code/do_git_checkout_new_branch.py
@@ -9,7 +9,7 @@ import argparse
 
 from z0lib import z0lib
 
-__version__ = "2.0.10"
+__version__ = "2.0.11"
 
 
 class RepoCheckout(object):