+++ wok_code/gen_addons_table.py
@@ -32,7 +32,7 @@ standard_library.install_aliases()  # noqa: E402
 
 MARKERS = r'(\[//\]: # \(addons\))|(\[//\]: # \(end addons\))'
 MANIFESTS = ('__openerp__.py', '__manifest__.py')
-__version__ = "2.0.10"
+__version__ = "2.0.11"
 
 
 class UserError(Exception):