+++ wok_code/scripts/odoo_dependencies.py
@@ -109,7 +109,7 @@ try:
 except ImportError:
     import clodoo
 
-__version__ = '2.0.10'
+__version__ = '2.0.11'
 
 
 MANIFEST_FILES = ['__manifest__.py', '__odoo__.py', '__openerp__.py', '__terp__.py']