+++ wok_code/scripts/please_cwd.py
@@ -6,7 +6,7 @@ import re
 
 import psycopg2
 
-__version__ = "2.0.10"
+__version__ = "2.0.11"
 
 
 class PleaseCwd(object):